Landmark-detection-Web-App/requirements.txt
streamlit
numpy
Pillow
pandas
geopy
tensorflow_hub
tensorflow==2.12.0
